AB#27294 show dynamic values in edit modal
