+++ applications/Unity.GrantManager/modules/Unity.Flex/src/Unity.Flex.Shared/Localization/Flex/en.json
@@ -31,6 +31,8 @@
     "Worksheet:Configuration:AddCheckboxOptionText": "Add Option",
     "Worksheet:Configuration:ExportWorksheetButtonText": "Export Worksheet",
     "Worksheet:Configuration:AddSelectListOptionText": "Add Option",
-    "Worksheet:Configuration:AddColumnOptionText": "Add Column"
+    "Worksheet:Configuration:AddColumnOptionText": "Add Column",
+    "DataGrids:DynamicColumnsHeader": "Dynamic Columns",
+    "DataGrids:CustomColumnsHeader": "Custom Columns"
   }
 }
