+++ applications/Unity.GrantManager/modules/Unity.Flex/src/Unity.Flex.Web/Pages/Components/DataGrid/DataGridReadService.cs
@@ -42,16 +42,16 @@ namespace Unity.Flex.Web.Pages.Flex
             return formattedKeyValuePairs;
         }
 
-        internal async Task<List<WorksheetFieldViewModel>?> GetPropertiesAsync(RowInputData dataProps)
+        internal async Task<(KeyValuePair<string, string>[] dynamicFields, List<WorksheetFieldViewModel>? customFields)> GetPropertiesAsync(RowInputData dataProps)
         {
             if (IsFirstRow(dataProps))
             {
-                return await GetFirstRowAsync(dataProps);
+                return ([], await GetFirstRowAsync(dataProps));
             }
 
             if (AddNewRow(dataProps))
             {
-                return await GetNewRowAsync(dataProps);
+                return ([], await GetNewRowAsync(dataProps));
             }
 
             return await GetExistingRowAsync(dataProps);
@@ -62,7 +62,7 @@ namespace Unity.Flex.Web.Pages.Flex
             var datagridDefinition = await customFieldAppService.GetAsync(dataProps.FieldId);
             if (datagridDefinition == null) return [];
 
-            return DataGridServiceUtils.ConvertDataGridValue(new DataGridValue(), datagridDefinition, 0, true);
+            return DataGridServiceUtils.ExtractCustomColumnsValues(new DataGridValue(), datagridDefinition, 0, true);
         }
 
         private async Task<List<WorksheetFieldViewModel>> GetNewRowAsync(RowInputData dataProps)
@@ -76,21 +76,22 @@ namespace Unity.Flex.Web.Pages.Flex
 
             var dataGridValue = JsonSerializer.Deserialize<DataGridValue>(customFieldValue.CurrentValue ?? "{}");
 
-            return DataGridServiceUtils.ConvertDataGridValue(dataGridValue, datagridDefinition, dataProps.Row, true);
+            return DataGridServiceUtils.ExtractCustomColumnsValues(dataGridValue, datagridDefinition, dataProps.Row, true);
         }
 
-        private async Task<List<WorksheetFieldViewModel>> GetExistingRowAsync(RowInputData dataProps)
+        private async Task<(KeyValuePair<string, string>[] dynamicFields, List<WorksheetFieldViewModel> customFields)> GetExistingRowAsync(RowInputData dataProps)
         {
             if (dataProps.ValueId == null) throw new ArgumentNullException(nameof(dataProps));
             var customFieldValue = await customFieldValueAppService.GetAsync(dataProps.ValueId.Value);
-            if (customFieldValue.CurrentValue == null) return [];
+            if (customFieldValue.CurrentValue == null) return ([], []);
 
             var datagridDefinition = await customFieldAppService.GetAsync(dataProps.FieldId);
-            if (datagridDefinition == null) return [];
+            if (datagridDefinition == null) return ([], []);
 
             var dataGridValue = JsonSerializer.Deserialize<DataGridValue>(customFieldValue.CurrentValue ?? "{}");
 
-            return DataGridServiceUtils.ConvertDataGridValue(dataGridValue, datagridDefinition, dataProps.Row, false);
+            return (DataGridServiceUtils.ExtractDynamicColumnsPairs(dataGridValue, dataProps.Row),
+                DataGridServiceUtils.ExtractCustomColumnsValues(dataGridValue, datagridDefinition, dataProps.Row, false));
         }
 
         private static bool AddNewRow(RowInputData dataProps)
