+++ applications/Unity.GrantManager/modules/Unity.Flex/src/Unity.Flex.Web/Views/Shared/Components/DataGridWidget/Default.js
@@ -170,6 +170,7 @@ $(function () {
         row,
         isNew,
         uiAnchor) {
+
         let formVersionId = $('#ApplicationFormVersionId').val();
         let applicationId = $('#DetailsViewApplicationId').val();
 
@@ -257,9 +258,11 @@ $(function () {
     function configureTable(table, fieldId) {
         table.buttons().container().prependTo(`#btn-container-${fieldId}`);
         let knownColumns = [];
+
         table.columns().header().to$().each(function (index) {
-            if ($(this).text() != '') {
-                knownColumns.push({ title: $(this).text(), visible: true, index: index + 1 });
+            let isActions = $(this).hasClass('custom-actions-header');
+            if ($(this).text() != '' && !isActions) {
+                knownColumns.push({ title: $(this).text(), visible: true, index: index + 1, isActions: isActions });
             }
         });
 
@@ -317,7 +320,8 @@ $(function () {
             tableDataSet.wsiId,
             rowDataSet.rowNo,
             false,
-            tableDataSet.uiAnchor);
+            tableDataSet.uiAnchor,
+            ['field1', 'field2']);
     }
 
     PubSub.subscribe(
